+++ client/src/App.js
@@ -9,6 +9,7 @@ import SubmissionFormPage from './pages/SubmissionFormPage'
 
 import { Route, Routes } from 'react-router-dom'
 
+
 function App() {
   return (
     <div className="container m-auto">