+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 - [] clear vendors and add some kind of validation of email, instagram names etc
 - [] front end - have instagram input automagically parse username on leaving focus?
 - [] implement form submission mutation
+- [] add client input to form submission
+- [] add event date to form submission
 
 ## 2023-04-09
 - updated submission form to allow adding and deleting of rows